refactor(#181): 스페이스 패키지 분리
